#387 UI for Rewards has been added

+++ domain/src/main/java/io/golos/domain/dto/RewardValueDomain.kt
@@ -4,6 +4,6 @@ package io.golos.domain.dto
  * Value of reward
  */
 data class RewardValueDomain(
-    val value: Float,
+    val value: Double,
     val communityId: String
 )
